requirements.txt
ultralytics>=8.3.0
opencv-python>=4.9.0.80
numpy>=1.26.0
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.1
pandas>=2.2.0
pytest>=8.0.0
httpx>=0.27.0
openai>=1.40.0
google-genai>=0.3.0

